Completed Functionality of Open Price Tab
