+++ src/components/open-price/InputDiv.jsx
@@ -1,4 +1,4 @@
-import OpenPriceInput from "./OpenPriceInput"
+
 
 export default function InputDiv({inputKey,openCount,inputEP,inputQuantity,handleInputEP,handleInputQuantity,handleDeleteEntry}){
   return (
