+++ src/components/open-price/OpenPriceResult.jsx
@@ -1,3 +1,4 @@
+import { calculateOpenPRice } from "../../functions/calculator";
 import Output from "../Output";
 import { useState, useEffect } from "react";
 
@@ -9,14 +10,20 @@ export default function OpenPriceResult({ valueArray }){
     
   ]);
 
- 
+  useEffect(() => {
+    const openPrice = calculateOpenPRice(valueArray);
+    const updatedProps = props.map((prop, index) => {
+      return { ...prop, Value: openPrice};
+    });
+    setProps(updatedProps);
+  },[valueArray])
 
   return (
     <>
       <div className="container-openPrice-output bg-blue-box">
         <span className="output-heading">Result</span>
         {props.map((prop) => (
-          <Output Class={prop.Class} Label={prop.Label} Value={prop.Value} Sign={prop.Sign} Id={prop.Id} key={prop.Id} />
+          <Output Class={prop.Class} Label={prop.Label} Value={prop.Value} Sign={prop.Sign} Id={prop.Id}/>
         ))}
       </div>
     </>
